Move table model into separate mod
